[FOLD] Address non-name-changing comments from @seelabs
