+++ src/ripple/consensus/LedgerTrie.h
@@ -21,7 +21,6 @@
 #define RIPPLE_APP_CONSENSUS_LEDGERS_TRIE_H_INCLUDED
 
 #include <ripple/basics/ToString.h>
-#include <ripple/basics/tagged_integer.h>
 #include <ripple/json/json_value.h>
 #include <algorithm>
 #include <memory>
