+++ src/ripple/protocol/impl/Indexes.cpp
@@ -21,10 +21,8 @@
 #include <ripple/protocol/SeqProxy.h>
 #include <ripple/protocol/digest.h>
 #include <ripple/protocol/nftPageMask.h>
-#include <boost/endian/conversion.hpp>
 #include <algorithm>
 #include <cassert>
-#include <string_view>
 
 namespace ripple {
